changed the visibilty of all classes to internal. moved all the function, tokens, and extensions from one long file to individual files because it doesnt make sense to have a file with 25+ classes
