+++ TastyScript/Lang/Extensions/ExtensionColor.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TastyScript.Lang.Extensions
+{
+    [Extension("Color", new string[] { "color" })]
+    [Serializable]
+    internal class ExtensionColor : EDefinition { }
+}
