+++ TastyScript/Lang/Extensions/ExtensionFor.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TastyScript.Lang.Extensions
+{
+    [Extension("For", new string[] { "enumerator" })]
+    [Serializable]
+    internal class ExtensionFor : EDefinition { }
+}
