+++ TastyScript/Lang/Extensions/ExtensionStart.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TastyScript.Lang.Extensions
+{
+    [Extension("Start"), Serializable]
+    internal class ExtensionStart : EDefinition { }
+}
