+++ TastyScript/Lang/Extensions/ExtensionStop.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TastyScript.Lang.Extensions
+{
+    [Extension("Stop"), Serializable]
+    internal class ExtensionStop : EDefinition { }
+}
