+++ TastyScript/Lang/Extensions/ExtensionThreshold.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TastyScript.Lang.Extensions
+{
+    [Extension("Threshold", new string[] { "int" })]
+    [Serializable]
+    internal class ExtensionThreshold : EDefinition { }
+}
