+++ TastyScript/Lang/Functions/FunctionBase.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using TastyScript.Lang.Extensions;
+using TastyScript.Lang.Token;
+
+namespace TastyScript.Lang.Functions
+{
+    [Function("Base", isSealed: true)]
+    internal class FunctionBase : FDefinition<object>
+    {
+        public override void TryParse(TParameter args, IBaseFunction caller, string lineval = "{0}")
+        {
+            Compiler.ExceptionListener.Throw(new ExceptionHandler(ExceptionType.CompilerException, $"{this.Name} can not be overrided", LineValue));
+        }
+        public override object CallBase(TParameter args)
+        {
+            Compiler.ExceptionListener.Throw(new ExceptionHandler(ExceptionType.CompilerException, $"{this.Name} can not be overrided.", LineValue));
+            return null;
+        }
+        protected override void ForExtension(TParameter args, ExtensionFor findFor, string lineval)
+        {
+            Compiler.ExceptionListener.Throw(new ExceptionHandler(ExceptionType.CompilerException, $"Cannot call 'For' on {this.Name}.", LineValue));
+        }
+    }
+}
