+++ TastyScript/TastyScript.csproj
@@ -107,12 +107,49 @@
     <Compile Include="Lang\Compiler.cs" />
     <Compile Include="Lang\Exceptions\ExceptionHandler.cs" />
     <Compile Include="Lang\Exceptions\IExceptionListener.cs" />
-    <Compile Include="Lang\Function\ExtensionDefinitions.cs" />
-    <Compile Include="Lang\Function\FunctionDefinitions.cs" />
-    <Compile Include="Lang\Function\IFunction.cs" />
-    <Compile Include="Lang\Function\Line.cs" />
+    <Compile Include="Lang\Extensions\ExtensionAnd.cs" />
+    <Compile Include="Lang\Extensions\ExtensionColor.cs" />
+    <Compile Include="Lang\Extensions\ExtensionConcat.cs" />
+    <Compile Include="Lang\Extensions\ExtensionElse.cs" />
+    <Compile Include="Lang\Extensions\ExtensionFor.cs" />
+    <Compile Include="Lang\Extensions\ExtensionOr.cs" />
+    <Compile Include="Lang\Extensions\ExtensionPrint.cs" />
+    <Compile Include="Lang\Extensions\ExtensionStart.cs" />
+    <Compile Include="Lang\Extensions\ExtensionStop.cs" />
+    <Compile Include="Lang\Extensions\ExtensionThen.cs" />
+    <Compile Include="Lang\Extensions\ExtensionThreshold.cs" />
+    <Compile Include="Lang\Functions\FunctionAppPackage.cs" />
+    <Compile Include="Lang\Functions\FunctionAwake.cs" />
+    <Compile Include="Lang\Functions\FunctionBase.cs" />
+    <Compile Include="Lang\Functions\FunctionBreak.cs" />
+    <Compile Include="Lang\Functions\FunctionCheckScreen.cs" />
+    <Compile Include="Lang\Functions\FunctionConnectDevice.cs" />
+    <Compile Include="Lang\Functions\FunctionContinue.cs" />
+    <Compile Include="Lang\Functions\FunctionGuaranteedHalt.cs" />
+    <Compile Include="Lang\Functions\FunctionHalt.cs" />
+    <Compile Include="Lang\Functions\FunctionIf.cs" />
+    <Compile Include="Lang\Functions\FunctionKeyEvent.cs" />
+    <Compile Include="Lang\Functions\FunctionLongTouch.cs" />
+    <Compile Include="Lang\Functions\FunctionLoop.cs" />
+    <Compile Include="Lang\Functions\FunctionPrint.cs" />
+    <Compile Include="Lang\Functions\FunctionPrintLine.cs" />
+    <Compile Include="Lang\Functions\FunctionSendText.cs" />
+    <Compile Include="Lang\Functions\FunctionSetDefaultSleep.cs" />
+    <Compile Include="Lang\Functions\FunctionSleep.cs" />
+    <Compile Include="Lang\Functions\FunctionStart.cs" />
+    <Compile Include="Lang\Functions\FunctionSwipe.cs" />
+    <Compile Include="Lang\Functions\FunctionTakeScreenshot.cs" />
+    <Compile Include="Lang\Functions\FunctionTimer.cs" />
+    <Compile Include="Lang\Functions\FunctionTouch.cs" />
+    <Compile Include="Lang\Extensions\EDefinition.cs" />
+    <Compile Include="Lang\Functions\FDefinition.cs" />
+    <Compile Include="Lang\IFunction.cs" />
+    <Compile Include="Lang\Line.cs" />
     <Compile Include="Lang\TokenParser.cs" />
     <Compile Include="Lang\Token\IToken.cs" />
+    <Compile Include="Lang\Token\TFunction.cs" />
+    <Compile Include="Lang\Token\TObject.cs" />
+    <Compile Include="Lang\Token\TParameter.cs" />
     <Compile Include="Lang\Utilities.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
